Merged in TPI-75-Create-Recorded-Inspections (pull request #24)
TPI-75 - Create Record Inspection

* TPI-75: Create OnInspectionLog event in InspectionWindowBuilder.cs. This event will be triggered when user select Compliant / Non Compliant
TPI-75: Add condition if the first inspection had photo but the reinspection doesn’t we DO NOT change the photo flag in Inspections.cs.
TPI-75: Build the tests for invoking OnInspectionLog Event and keeping the photo status if  the reinspection doesn’t

* Merge branch 'main' into TPI-75-Create-Recorded-Inspections

* TPI-75: Removed methods in EventSystem
TPI-75: Replaced hasPhoto with photoTaken
TPI-75: Added a private InspectionData inspectionData to the properties list
TPI-75: Updated the RecordInspection() method to record the inspection data by assigning the current inspectable object, the compliance status and photo was taken.
TPI-75: Call the RecordInspection() in CompliantSelected() and NonCompliantSelected()

* TPI-75: Remove unused using statement : using UnityEngine.PlayerLoop

* TPI-75 remove unused statement

* TPI-75: Cleanup code

* TPI-75: Remove unused variable

* Merge branch 'main' into TPI-75-Create-Recorded-Inspections

# Conflicts:
#	Assets/Scenes/Main Scene.unity
#	Assets/Scripts/UI/InspectionWindowBuilder.cs
#	Assets/Tests/Play Mode/InspectionWindowIntegrationTest.cs


Approved-by: Kristin Theoret
Approved-by: Juhwan Seo
Approved-by: Carolina Naoum Junqueira