Merged in TPI-87-asset-import-24feb2025 (pull request #27)
TPI-87 asset import 24feb2025

* TPI-87 added the complex material script

* added the textures and created the materials for the ink bottles, model is going to be revised and will be imported later.

* TPI-87 imported tool box asset

* TPI-30 Tool Box
* imported textures and model, created material and created a tool box prefab

* TPI-87 - imported the tripod light assets

* TPI-31 ring light

* TPI-87 imported PPE gloves and glove box

* TPI-39 Sterile Gloves

* TPI-87 asset import

* TPI-47 waste basket

* TPI-87 asset import

*TPI-40 hand sanitizer

* TPI-87 asset import

* TPI-33 Salon Cart

* Merge branch 'main' into TPI-87-asset-import-24feb2025

* Made several adjustments:

Firstly, created some Presets that will look for imported files with the suffix of _normal, and _Normal and automatically tag those images as normal maps
Related to above, went through all normal map texture files in the project and set them to be Normal Maps

Adjusted the ShaderGraph for the MRO Triple Mask sot aht "Has Rim Light" is no longer on by default
Related to the above, went through all new imports and disabled the "Has Rim Light" option on all materials for this import

Created an alpa variation of the Trach Can material so that we can see the mesh, and applied it to the prefab

Adjusted the ToolChest material to use a CVM so it can be dyed any colour we want
Similarly, dyed the Tool Chest red, and the gloves purple.

Adjusted the hierachy of the Tripod Light assets so that they inherent transforms logically - moving the middle section will now also move the top and light, but not the bottom, etc.


Approved-by: Topher Rouleau